refactor(utils): consolidate util and reauth into utils submodule
rename util.py → utils/crypto.py (content is HMAC-SHA1 only)
move reauth.py → utils/reauth.py

Groups all shared helpers alongside utils/redact.py, clears
top-level namespace.
